// apps/api/src/modules/click2mail/interfaces/IClick2mailRequestCreateAddress.ts
export interface IClick2mailRequestCreateAddress {
  suffix: string;
  lastName: string;
  middleName: string;
  firstName: string;
  prefix: string;
  address1: string;
  address2: string;
  address3: string;
  city: string;
  state: string;
  zip: string;
  country: string;
  description: string;
  makeDefault: string;
  type: string;
  phone: string;
}